fix callrecord

+++ Model/DTO/AddCallRecord.cs
@@ -4,6 +4,6 @@
     {
         public string UserId { get; set; }
         public string CamgirlId { get; set; }
-        public int TimeUsed { get; set; }
+        public decimal TimeUsed { get; set; }
     }
 }
